requirements.txt
aniso8601==9.0.1
attrs==21.4.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pycparser==2.21
PyJWT==1.5.3
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
